feat: add AbortSignal support at the task level
Fixes #139

This change enables `Task` to be aborted via an `AbortSignal` which falls back to a signal its
`Bench`'s signal.

It adds a `signal` prop to the `Task` class, as well as an `isAborted()` method (which could be a
getter, if you wish).

I'm wrapping `tinybench` and have confirmed this works well, as I'm able to abort (async) tasks via
the signal.

I've added tests covering intended usage and some edge cases, but it seems there may be a
memory-leak issue with concurrent tasks (I'm not sure if this is a known bug); this test
(`task-level abort: works with task concurrency`) is skipped.  If you can get it working, then
maybe it's just my environment.

I added documentation about the bench-and-task-level-`AbortSignal` support to `README.md`.  Happy
to remove if it's not needed--or I can remove the examples, or whatever.